добавлен функционал ручного проставления цен при регистрации билета

+++ site/templates/editticket.php
@@ -17,21 +17,24 @@ $old_pay_or_booking = !empty($_POST['old_pay_or_booking'])?$_POST['old_pay_or_bo
 $old_booking_sum = !empty($_POST['old_booking_sum'])?$_POST['old_booking_sum']:NULL;
 $old_confirm = !empty($_POST['old_confirm'])?$_POST['old_confirm']:NULL;
 $old_agent_ticket = !empty($_POST['old_agent_ticket'])?$_POST['old_agent_ticket']:NULL;
+$old_price_ticket = !empty($_POST['old_price_ticket'])?$_POST['old_price_ticket']:NULL;
 $passenger = !empty($_POST['passenger'])?$_POST['passenger']:NULL;
 
 $pay_or_booking = !empty($_POST['pay_or_booking'])?$_POST['pay_or_booking']:NULL;
 $booking_sum = !empty($_POST['booking_sum'])?$_POST['booking_sum']:NULL;
 $confirm = !empty($_POST['confirm'])?$_POST['confirm']:NULL;
 $agent_ticket = !empty($_POST['agent_ticket'])?$_POST['agent_ticket']:NULL;
+$price_ticket = !empty($_POST['price_ticket'])?$_POST['price_ticket']:NULL;
 
 $success = 'Статус билета успешно изменен';
 $log = '';
-if ($selected_bus && $selected_id_bus && $selected_date && $selected_time && $selected_seat && $id_seat && $old_pay_or_booking && $old_confirm && $old_agent_ticket && $passenger && $pay_or_booking && $confirm && $agent_ticket) {
-    $log .= date('Y-m-d H:i:s') . ' - Изменен статус в билете id - ' . $id_seat . '. ';
+if ($selected_bus && $selected_id_bus && $selected_date && $selected_time && $selected_seat && $id_seat && $old_pay_or_booking && $old_confirm && $old_agent_ticket && $passenger && $pay_or_booking && $confirm && $agent_ticket && $price_ticket) {
+    $log .= date('Y-m-d H:i:s') . ' - Изменен билет id - ' . $id_seat . '. ';
     $log .= 'Статус изменен с ' . $old_pay_or_booking . ' на ' . $pay_or_booking . ' оператором ' . $operator . '. '; 
     $log .= 'Сумма к оплате при бронировании изменена с ' . $old_booking_sum . ' на ' . $booking_sum . ' оператором ' . $operator . '. ';
     $log .= 'Статус подтверждения изменен с ' . $old_confirm . ' на ' . $confirm . ' оператором ' . $operator . '. ';  
     $log .= 'Агент изменен с ' . $old_agent_ticket . ' на ' . $agent_ticket . ' оператором ' . $operator . '. ';
+    $log .= 'Цена билета изменена с ' . $old_price_ticket . ' на ' . $price_ticket . ' оператором ' . $operator . '. ';
     $log .= 'Параметры измененного билета: ' . $selected_bus . ' ' . $selected_date . '' . $selected_time . ', id автобуса - ' . $selected_id_bus . ', место - ' . $selected_seat . ', пассажир - ' . $passenger; 
     file_put_contents(__DIR__ . '/log_edit_tikets.txt', $log . PHP_EOL, FILE_APPEND);
     
@@ -40,8 +43,8 @@ if ($selected_bus && $selected_id_bus && $selected_date && $selected_time && $se
     $edit_page->pay_or_booking = $pay_or_booking;
     $edit_page->booking_sum = $booking_sum;
     $edit_page->confirm = $confirm;
-    $edit_page->confirm = $confirm;
     $edit_page->agent_ticket = $agent_ticket;
+    $edit_page->price_ticket = $price_ticket;
     $edit_page->save();
 } else {
     $success = 'Статус билета не изменен!<br>Ошибка в данных';
